Add ability to respond to offset fetch requests
This is a very naive and slow implementation since it pulls down
basically all logs of a topic-partition to scan through and find the
ones corresponding to offset requests for the given (consumer group id,
topic, partition).

A more sophisticated way to do this would be to have a thread that
independently follows the log and keeps track of all consumer offsets.
Then, reading the offsets would involve just doing a cache lookup.